refactor(docker): move to separate solution folder
