+++ src/Thinktecture.Relay.Server.Abstractions/ITenantResolver.cs
@@ -11,7 +11,7 @@ namespace Thinktecture.Relay.Server.Abstractions
 		/// Resolves a tenant identity to a tenant.
 		/// </summary>
 		/// <param name="identity">The identity of the tenant.</param>
-		/// <returns>A <see cref="Task"/> representing the asynchronous operation, which wraps the tenant or <value>null</value>
+		/// <returns>A <see cref="Task"/> representing the asynchronous operation, which wraps the tenant or null
 		/// if the identity is unknown.</returns>
 		Task<object> ResolveAsync(object identity);
 		// TODO replace with model and identity
