--- src/Thinktecture.Relay.Server.Persistence.EntityFrameworkCore/ApplicationBuilderExtensions.cs
@@ -1,10 +1,0 @@
-// ReSharper disable once CheckNamespace; (extension methods on IApplicationBuilder namespace)
-namespace Microsoft.AspNetCore.Builder
-{
-	/// <summary>
-	/// Extension methods for the <see cref="IApplicationBuilder"/>.
-	/// </summary>
-	public static class ApplicationBuilderExtensions
-	{
-	}
-}
