+++ src/docker/Thinktecture.Relay.Docker/Program.cs
@@ -1,0 +1,17 @@
+using System;
+using System.Threading.Tasks;
+using Microsoft.Extensions.Hosting;
+using Serilog;
+using Thinktecture.Relay.Docker;
+
+namespace Thinktecture.Relay.Docker
+{
+	public class Program
+	{
+		public static int Main()
+		{
+			Console.WriteLine("This cannot be run.");
+			return 1;
+		}
+	}
+}
