+++ src/docker/Thinktecture.Relay.IdentityServer.Docker/Dockerfile
