+++ src/docker/Thinktecture.Relay.IdentityServer.Docker/Properties/launchSettings.json
