+++ src/docker/Thinktecture.Relay.IdentityServer.Docker/appsettings.Development.json
