+++ src/docker/Thinktecture.Relay.IdentityServer.Docker/appsettings.json
