+++ src/docker/Thinktecture.Relay.IdentityServer.Docker/wwwroot/favicon.ico
