+++ src/docker/Thinktecture.Relay.Server.Docker/Controllers/HealthController.cs
