+++ src/docker/Thinktecture.Relay.Server.Docker/Dockerfile
