+++ src/docker/Thinktecture.Relay.Server.Docker/Properties/launchSettings.json
