+++ src/docker/Thinktecture.Relay.Server.Docker/appsettings.json
