+++ src/docker/Thinktecture.Relay.Server.Docker/run-container.ps1
